requirements.txt
pyodbc==5.3.0
azure-communication-email==1.1.0
